refactor: reorganize composables and remove unused files in sidepanel
