+++ packages/next-wxt/entrypoints/sidepanel/App.vue
@@ -1,17 +1,17 @@
 <script lang="ts" setup>
 import { ref, type Ref, shallowReactive, computed, watch } from 'vue'
 import { TinyRemoter } from '@opentiny/next-remoter'
-import { useBrowserExtensions } from './useBrowserExtensions'
-import { useWebAgentServer } from './useWebAgentServer'
+import { useBrowserExtensions } from './composable/useBrowserExtensions'
+import { useWebAgentServer } from './composable/useWebAgentServer'
 import TinyUser from '@opentiny/vue-user'
-import { useCustomMarketMcpServers } from './useCustomMarketMcpServers'
+import { useCustomMarketMcpServers } from './composable/useCustomMarketMcpServers'
 import { TrSuggestionPillButton, TrDropdownMenu } from '@opentiny/tiny-robot'
 import { AGENT_ROOT, ROBOT_URL } from './const'
-import { useGenerateCode } from './useGenerateCode'
+import { useGenerateCode } from './composable/useGenerateCode'
 import RecordModal from './components/RecordModal.vue'
 import { getAllSkills } from '@/skills'
 import { RENDERER_SETTINGS_KEY } from '@opentiny/genui-sdk-vue'
-import { useAutoScreenshot } from './useAutoScreenshot'
+import { useAutoScreenshot } from './composable/useAutoScreenshot'
 import { CustomFunction } from '@/utils/customFunction'
 import { DEFAULT_MODEL_CONFIGS } from './model-config'
 import { getStorageItem, setStorageItem } from './utils/local-storage'
